[refactor] util 디렉토리를 infrastructure로 이동
