+++ backend/src/main/java/com/ddbb/dingdong/application/usecase/auth/CheckEmailUseCase.java
@@ -3,7 +3,7 @@ package com.ddbb.dingdong.application.usecase.auth;
 import com.ddbb.dingdong.application.common.Params;
 import com.ddbb.dingdong.application.common.UseCase;
 import com.ddbb.dingdong.application.usecase.auth.errors.AuthInvalidParamErrors;
-import com.ddbb.dingdong.util.ParamValidator;
+import com.ddbb.dingdong.infrastructure.util.ParamValidator;
 import com.ddbb.dingdong.domain.auth.service.AuthManagement;
 import com.ddbb.dingdong.presentation.endpoint.auth.exchanges.CheckEmailDto;
 import lombok.AllArgsConstructor;
