+++ backend/src/main/java/com/ddbb/dingdong/application/usecase/user/TestFCMNotificationUseCase.java
@@ -25,5 +25,4 @@ public class TestFCMNotificationUseCase implements UseCase<TestFCMNotificationUs
     public static class Param implements Params {
         private Long userId;
     }
-
 }
