+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/bus/subscription/subscriber/SocketSubscriber.java
@@ -1,7 +1,7 @@
 package com.ddbb.dingdong.infrastructure.bus.subscription.subscriber;
 
 import com.ddbb.dingdong.infrastructure.bus.subscription.BusSubscriptionManager;
-import com.ddbb.dingdong.util.FormatUtil;
+import com.ddbb.dingdong.infrastructure.util.FormatUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.geo.Point;
 import org.springframework.web.socket.TextMessage;
