+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/routing/api/TmapApiProperties.java
@@ -5,13 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
 @Setter
 @Component
 @ConfigurationProperties(prefix = "api.tmap")
