+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/util/ParamValidator.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.util;
+package com.ddbb.dingdong.infrastructure.util;
 
 import java.util.regex.Pattern;
